Keep all in SharedTreeModel as it contribs are only utilized by drf or gbm
